+++ src/Components/Page/IBAS/Activity/Activity.tsx
@@ -2,93 +2,27 @@ import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import useFetch from "../../../../Hooks/useFetch";
 
-import styled from "styled-components";
 import Button from "../../../../styles/assets/Button";
 import { Div, FlexDiv } from "../../../../styles/assets/Div";
 import Img from "../../../../styles/assets/Img";
 import P from "../../../../styles/assets/P";
 
 import Pagination from "../../../Common/Pagination";
+import ActivityCard from "../../../Component/IBAS/Activity/ActivityCard";
 
 import { useRecoilState } from "recoil";
 import { totalPageInfo } from "../../../../Recoil/backState";
 
-const Article = styled.article`
-    position: relative;
-    width: 360px;
-    height: 360px;
-`;
+import { GetRoleAuthorization } from "../../../../Functions/authFunctions";
 
-const ArticleImg = styled(Img)`
-    filter: brightness(60%);
-    cursor: pointer;
-    &:hover {
-        filter: brightness(70%);
-    }
-`;
-
-export interface activityInterface {
-    id: number;
-    title: string;
-    writerName: string;
-    dateCreated: string;
-    dateUpdated: string;
-    thumbnail: {
-        id: string;
-        name: string;
-        url: string;
-        size: number;
-        type: string;
-    } | null;
-}
-
-const ActivityComponent = ({ imgSrc, title, dateCreated, writerName, id }: any) => {
-
-    const navigate = useNavigate();
-
-    const moveDetail = () => navigate(`/activity/detail/${id}`);
-
-    return (
-        <>
-            <Article onClick={() => moveDetail()}>
-                <Div width="360px" height="360px" overflow="hidden" radius={10}>
-                    <ArticleImg src={imgSrc} />
-                </Div>
-                <Div $position="absolute" $bottom="20px" $left="10px" $padding="0 10px">
-                    <Div $margin="0 0 15px 0">
-                        <P color="wh" fontSize="xl" fontWeight={800}>
-                            {title}
-                        </P>
-                    </Div>
-                    <FlexDiv>
-                        <FlexDiv width="12px" $margin="0 3px 0 0">
-                            <Img src="/images/user_white.svg" />
-                        </FlexDiv>
-                        <Div>
-                            <P color="wh" fontSize="sm">
-                                {writerName}
-                            </P>
-                        </Div>
-                        <FlexDiv width="12px" $margin="0 5px ">
-                            <Img src="/images/clock_white.svg" />
-                        </FlexDiv>
-                        <Div>
-                            <P color="wh" fontSize="sm">
-                                {dateCreated?.split('T')[0]} {dateCreated?.split('T')[1]}
-                            </P>
-                        </Div>
-                    </FlexDiv>
-                </Div>
-            </Article>
-        </>
-    )
-}
+import { ActivityInterface } from "../../../../Types/IBAS/TypeIBAS";
 
 const Activity = () => {
     const navigate = useNavigate();
     const [totalPage, setTotalPage] = useRecoilState(totalPageInfo);
     const [activityListData, fetchActivityListData] = useFetch();
-    const [detail, setDetail] = useState<activityInterface[] | null>(null);
+    const [detail, setDetail] = useState<ActivityInterface[] | null>(null);
+    const { isAuthorizedOverSecretary } = GetRoleAuthorization();
 
     useEffect(() => {
         fetchActivityListData('/club/activities?page=0&size=6', 'GET', 'token')
@@ -109,7 +43,7 @@ const Activity = () => {
                 <FlexDiv width="75%" $justifycontent="space-between" $margin="50px 0 0 0">
                     {detail?.slice(0, 3)?.map(({thumbnail, title, dateCreated, writerName, id}) => (
                         <Div width="25%">
-                            <ActivityComponent
+                            <ActivityCard
                             imgSrc = {thumbnail?.url}
                             title = {title}
                             dateCreated = {dateCreated}
@@ -123,7 +57,7 @@ const Activity = () => {
                 <FlexDiv width="75%" $justifycontent="space-between" $margin="50px 0 0 0">
                     {detail?.slice(3, 6)?.map(({thumbnail, title, dateCreated, writerName, id}) => (
                         <Div width="25%">
-                            <ActivityComponent
+                            <ActivityCard
                             imgSrc = {thumbnail?.url}
                             title = {title}
                             dateCreated = {dateCreated}
@@ -142,25 +76,27 @@ const Activity = () => {
                         size={6}
                     />
                 </Div>
-                <FlexDiv $margin="50px 0" width="75%" $justifycontent="end">
-                    <Button
-                        display="flex"
-                        $backgroundColor="bgColor"
-                        $padding="12px 15px"
-                        $borderRadius={30}
-                        $HBackgroundColor="bgColorHo"
-                        onClick={() => moveCreate()}
-                    >
-                        <Div width="12px" $margin="0 10px 0 0">
-                            <Img src="/images/plus_white.svg" />
-                        </Div>
-                        <Div $pointer>
-                            <P color="wh" fontSize="sm">
-                                게시글 등록
-                            </P>
-                        </Div>
-                    </Button>
-                </FlexDiv>
+                { isAuthorizedOverSecretary && (
+                    <FlexDiv $margin="50px 0" width="75%" $justifycontent="end">
+                        <Button
+                            display="flex"
+                            $backgroundColor="bgColor"
+                            $padding="12px 15px"
+                            $borderRadius={30}
+                            $HBackgroundColor="bgColorHo"
+                            onClick={() => moveCreate()}
+                        >
+                            <Div width="12px" $margin="0 10px 0 0">
+                                <Img src="/images/plus_white.svg" />
+                            </Div>
+                            <Div $pointer>
+                                <P color="wh" fontSize="sm">
+                                    게시글 등록
+                                </P>
+                            </Div>
+                        </Button>
+                    </FlexDiv>
+                )}
             </FlexDiv>
         </>
     );
