+++ src/Components/Page/IBAS/Activity/ActivityCreate.tsx
@@ -12,12 +12,13 @@ import DragNDrop from "../../../Common/DragNDrop";
 import TextEditor from "../../../Common/TextEditor";
 import Loading from "../../../Common/Loading";
 
+import { ActivityDetailInterface } from "../../../../Types/IBAS/TypeIBAS";
+
 import useFetch from "../../../../Hooks/useFetch";
 import { useSetRecoilState, useRecoilState, useRecoilValue } from "recoil";
 
 import { selectedFile, menuId, refetch } from "../../../../Recoil/frontState";
 import { fileIdList } from "../../../../Recoil/backState";
-
 
 const TitleTextInput = styled(TextInput)`
     border-radius: 5px;
@@ -27,22 +28,6 @@ const TitleTextInput = styled(TextInput)`
         color: ${(props) => props.theme.color.grey1};
     }
 `;
-
-export interface activityDetailInterface {
-    id: string;
-    title: string;
-    content: string;
-    writerName: string;
-    dateCreated: string;
-    dateUpdated: string;
-    images: {
-            id: string;
-            name: string;
-            url: string;
-            size: number;
-            type: string;
-    }[];
-}
 
 const ActivityCreate = () => {
     const navigate = useNavigate();
@@ -55,7 +40,7 @@ const ActivityCreate = () => {
     const setSelectedFile = useSetRecoilState(selectedFile);
     const [fileId, setFileList] = useRecoilState(fileIdList);
     const [files, setFiles] = useRecoilState(selectedFile);
-    const [detail, setDetail] = useState<activityDetailInterface | null>(null);
+    const [detail, setDetail] = useState<ActivityDetailInterface | null>(null);
     const currentMenuId = useRecoilValue(menuId);
     const setReload = useSetRecoilState(refetch);
 
@@ -83,9 +68,7 @@ const ActivityCreate = () => {
             formdata.append("files", files[i]);
         }
 
-        console.log(inputData)
         if (update === 'update') {
-            console.log(inputData)
             postFetchData(`/club/activity/${paramID}`, "POST", "token", inputData);
         } else {
             postFetchData(`/club/activity`, "POST", "token", inputData);
@@ -120,18 +103,17 @@ const ActivityCreate = () => {
     useEffect(() => {
         if (getData) {
             setDetail(getData);
-            console.log(detail)
             setIsLoading(false);
             
             // DragNDrop update 설정
             const files = [
-                ...getData.images.map((item: activityDetailInterface|null) => item),
-                ...getData.otherFiles.map((item: activityDetailInterface|null) => item),
+                ...getData.images.map((item: ActivityDetailInterface|null) => item),
+                ...getData.otherFiles.map((item: ActivityDetailInterface|null) => item),
             ];
             setSelectedFile(files);
             const fileIds = [
-                ...getData.images.map((item: activityDetailInterface|null) => item?.id),
-                ...getData.otherFiles.map((item: activityDetailInterface|null) => item?.id),
+                ...getData.images.map((item: ActivityDetailInterface|null) => item?.id),
+                ...getData.otherFiles.map((item: ActivityDetailInterface|null) => item?.id),
             ];
             setFileList(fileIds);
             setReload(true);
